Querying data - high level flow
This is high level flow where the query converted to vector by llamaindex and send to vector store to find the top k closest matches, which will serve as a context to LLM in addition to query to get the answers. As of now, this is done in a single call. Next, we will deep dive into it and find out what's happening in behind the scenes.